changes: organization of entities path
