+++ backend/src/domain/entities/appointment.ts
@@ -1,35 +1,35 @@
-import { User } from './user.entity';
-import { Work } from './work.entity';
-import { Business } from './business.entity';
+import { Work } from './work';
+import { User } from './user/user';
+import { Status } from '@/domain/constants/status.enum';
+import { Business, BusinessService } from './business';
 
-export class Payment {
+export class Appointment {
   public id: string;
-  public amount: number;
-  public status: string;
-  public paymentMethod: string;
-  public transactionId?: string;
-  public timestamp: Date;
-  public workId: string;
-  public work?: Work;
+  public datetime: Date;
+  public state: Status;
   public clientId: string;
   public client?: User;
+  public serviceId: string;
+  public service?: BusinessService;
   public businessId: string;
   public business?: Business;
+  public work?: Work;
+  public messages?: any[];
 
   /**
-   * Updates the payment with the provided data.
+   * Updates the appointment with the provided data.
    *
    * @param partial - A partial object containing the updated data. The following
    *                  properties are not allowed and will be deleted if present:
    *                    - id
-   *                    - workId
    *                    - clientId
+   *                    - serviceId
    *                    - businessId
    */
-  update(partial: Partial<Payment>) {
+  update(partial: Partial<Appointment>) {
     delete partial.id;
-    delete partial.workId;
     delete partial.clientId;
+    delete partial.serviceId;
     delete partial.businessId;
 
     Object.assign(this, partial);
