+++ backend/src/domain/entities/business/index.ts
@@ -1,0 +1,5 @@
+export * from './business-category';
+export * from './business-image';
+export * from './business-profile';
+export * from './business-service';
+export * from './business';
