+++ backend/src/domain/entities/index.ts
@@ -1,16 +1,10 @@
-export * from './account.entity';
-export * from './appointment.entity';
-export * from './availability.entity';
-export * from './business-category.entity';
-export * from './business-image.entity';
-export * from './business-profile.entity';
-export * from './business-service.entity';
-export * from './business.entity';
-export * from './category.entity';
-export * from './payment.entity';
-export * from './review.entity';
-export * from './token-version.entity';
-export * from './user-profile.entity';
-export * from './user.entity';
-export * from './verification.entity';
-export * from './work.entity';
+export * from './account';
+export * from './appointment';
+export * from './availability';
+export * from './business/business';
+export * from './category';
+export * from './payment';
+export * from './review';
+export * from './token-version';
+export * from './verification';
+export * from './work';
