+++ backend/src/domain/entities/user/index.ts
@@ -1,0 +1,2 @@
+export * from './user-profile';
+export * from './user';
