+++ backend/src/domain/interfaces/account-repository.ts
@@ -1,4 +1,4 @@
-import { User } from '../entities';
+import { User } from '../entities/user';
 
 export interface AccountRepository {
   findAccount(provider: string, providerId: string): Promise<User | null>;
